refactor: migrate to block portal path
